댓글 css 수정, 사용한 쿠폰 삭제, 결제 시 적립금 지급 등 코드 수정

+++ src/main/java/com/project/tour/domain/Coupon.java
@@ -18,6 +18,6 @@ public class Coupon {
 
     private String couponName;
 
-    private double couponRate;
+    private int couponRate;
 
 }
